openacc bug fixed

+++ src/print_measure.c
@@ -18,16 +18,16 @@ static int cmp_uint64 (const void *a, const void *b)
     return 0;
 }
 
-void print_measure(int n, int nrep, uint64_t tdiff[NB_META])
+void print_measure(int n, unsigned int nrep, uint64_t tdiff[NB_META])
 {
     FILE * output = NULL;
-
-    const unsigned int nbitr = n*n*n*nrep;
+    
+    const unsigned long nbitr = (unsigned long)n*(unsigned long)n*(unsigned long)n*(unsigned long)nrep;
 
     qsort (tdiff, NB_META, sizeof tdiff[0], cmp_uint64);
-    printf("Minimum : %.6g %s (%.3f par itération)\n", (float)tdiff[0]/nrep        , MEASURE_UNITE, (float)tdiff[0]/nbitr);
-    printf("Median  : %.6g %s (%.3f par itération)\n", (float)tdiff[NB_META/2]/nrep, MEASURE_UNITE, (float)tdiff[NB_META/2]/nbitr);
-    printf("Maximum : %.6g %s (%.3f par itération)\n", (float)tdiff[NB_META-1]/nrep, MEASURE_UNITE, (float)tdiff[NB_META-1]/nbitr);
+    printf("Minimum : %.6g %s (%.2g par itération)\n", (float)tdiff[0]/nrep        , MEASURE_UNITE, (float)tdiff[0]/nbitr);
+    printf("Median  : %.6g %s (%.2g par itération)\n", (float)tdiff[NB_META/2]/nrep, MEASURE_UNITE, (float)tdiff[NB_META/2]/nbitr);
+    printf("Maximum : %.6g %s (%.2g par itération)\n", (float)tdiff[NB_META-1]/nrep, MEASURE_UNITE, (float)tdiff[NB_META-1]/nbitr);
     
     const float stabilite = (tdiff[NB_META/2] - tdiff[0]) * 100.0f / tdiff[0];
